// src/types/auth.ts
export interface User {
    id: string;
    name: string;
    email: string;
    password?: string;
    role: string;
}

export interface LoginCredentials {
    email: string;
    password?: string;
}

export interface RegisterCredentials {
    name: string;
    email: string;
    password?: string;
}